Update app.py
updated Authentication

+++ orderapi/api_gateway/app.py
@@ -1,6 +1,6 @@
-from flask import Flask, request,jsonify, redirect, url_for
+from flask import Flask, request,jsonify, redirect, url_for,session
 import os
-
+from flask_jwt_extended import JWTManager, jwt_required, create_access_token
 from pymongo import MongoClient
 from datetime import datetime
 import sys
@@ -44,10 +44,32 @@ mycol5 = mydb["order"]
 
 app = Flask(__name__)
 
+#config JWT
+
+app.config["JWT_SECRET_KEY"] = "Comp3122Project" 
+app.config['SECRET_KEY'] = "Comp3122Project" 
+app.config['JWT_TOKEN_LOCATION'] = ['headers', 'query_string']
+jwt = JWTManager(app)
 @app.route('/', methods=['GET'])
 def index_page():
     #Index page   
-    return 'Order Shop Api\n', 200
+    return 'Order Shop Api \n '+'please login by /login/customer_id/password', 200
+
+@app.route('/login/<customer_id>/<password>', methods=['GET'])
+def login(customer_id,password):
+    test = mycol.find_one({"customer_id": customer_id,"password":password})
+    if test:
+        access_token = create_access_token(identity=customer_id)
+        session['token'] = access_token
+        return jsonify(message="Login Succeeded!", access_token=access_token), 201
+    else:
+        return jsonify(message="Bad customer ID or Password"), 401
+        
+@app.route('/logout', methods=['GET'])
+def logout():     
+        session.clear()
+        return jsonify(message="Logout Succeeded!"), 201
+
 
 @app.route('/me', methods=['GET'])
 def about_me():
@@ -55,12 +77,20 @@ def about_me():
     my_info = {
                 'id':'17077493D',
                 'name':'Lam Weidong'
-                }       
+                },{
+                'id':'21032101D',
+                'name':'Lau KwongHo'
+                }                        
     return jsonify(my_info)
+
+
 
 # Get all customers
 @app.route('/customers', methods=['GET'])
 def get_customers():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[]    
     results = list(mycol.find({},{"_id":0},).sort('customer_id'))
 
@@ -78,6 +108,9 @@ def get_customers():
 # Get customer by ID
 @app.route('/customers/<customer_id>', methods=['GET'])
 def get_customer_by_id(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[]    
     results = list(mycol.find({ "customer_id": customer_id},{'_id':0}))
     
@@ -91,6 +124,9 @@ def get_customer_by_id(customer_id):
 # get carts created by all customers
 @app.route('/cart', methods=['GET'])
 def get_carts():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[] 
     pipeline=[   
         { "$project": {'_id':0}},     
@@ -126,6 +162,9 @@ def get_carts():
 # get cart created by specified customer
 @app.route('/cart/<customer_id>', methods=['GET'])
 def get_cart_by_id(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output = []
     pipeline=[
         { "$match": {'customer_id':customer_id}},           
@@ -159,6 +198,9 @@ def get_cart_by_id(customer_id):
 #adding customer
 @app.route('/customers', methods=['POST'])
 def add_customer():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     data = request.json
     customer_id = data["customer_id"]
     username = data["username"]
@@ -170,6 +212,9 @@ def add_customer():
 #editing customer
 @app.route('/customers/<customer_id>', methods=['PATCH'])
 def edit_customer(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     new_data = request.json
     username = new_data["username"]
     deliver_address = new_data["deliver_address"]
@@ -180,6 +225,9 @@ def edit_customer(customer_id):
 #deleting customer
 @app.route('/customers/<customer_id>/delete', methods=['DELETE'])
 def del_customer(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     del_data = { "customer_id" : customer_id }
     mycol.delete_one(del_data)
     return jsonify({"message":"deleted customer"}),200
@@ -187,6 +235,9 @@ def del_customer(customer_id):
 #creating cart
 @app.route('/cart', methods=['POST'])
 def add_cart():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     data = request.json
     customer_id = data["customer_id"]
     item_ids = data["item_ids"]
@@ -197,6 +248,9 @@ def add_cart():
 #editing cart
 @app.route('/cart/<customer_id>', methods=['PATCH'])
 def edit_cart(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     new_data = request.json
     item_ids = new_data["item_ids"]
     shop_id = new_data["shop_id"]
@@ -206,6 +260,9 @@ def edit_cart(customer_id):
 #empty cart
 @app.route('/cart/<customer_id>/delete', methods=['DELETE'])
 def empty_cart(customer_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     del_data = { "customer_id" : customer_id }
     mycol2.delete_one(del_data)
     return jsonify({"message":"cart emptied"}),200
@@ -213,6 +270,9 @@ def empty_cart(customer_id):
 #Get all shops
 @app.route('/shops', methods=['GET'])
 def get_shops():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[]    
     results = list(mycol3.find({},{"_id":0},).sort('shop_id'))
 
@@ -230,6 +290,9 @@ def get_shops():
 # Get shop by ID
 @app.route('/shops/<shop_id>', methods=['GET'])
 def get_shop_by_id(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[]    
     results = list(mycol3.find({ "shop_id": shop_id},{'_id':0}))
     
@@ -243,6 +306,9 @@ def get_shop_by_id(shop_id):
 # get menu created by all shops
 @app.route('/menu', methods=['GET'])
 def get_menu():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[] 
     pipeline=[   
         { "$project": {'_id':0}},     
@@ -278,6 +344,9 @@ def get_menu():
 # get menu created by specified shop
 @app.route('/menu/<shop_id>', methods=['GET'])
 def get_menu_by_id(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     output =[] 
     pipeline=[
         { "$match": {'shop_id':shop_id}},           
@@ -311,6 +380,9 @@ def get_menu_by_id(shop_id):
 #adding shop
 @app.route('/shops', methods=['POST'])
 def add_shop():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     data = request.json
     shop_id = data["shop_id"]
     shopname = data["shopname"]
@@ -322,6 +394,9 @@ def add_shop():
 #editing shop
 @app.route('/shops/<shop_id>', methods=['PATCH'])
 def edit_shop(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     new_data = request.json
     shopname = new_data["shopname"]
     shop_address = new_data["shop_address"]
@@ -332,6 +407,9 @@ def edit_shop(shop_id):
 #deleting shop
 @app.route('/shops/<shop_id>/delete', methods=['DELETE'])
 def del_shop(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     del_data = { "shop_id" : shop_id }
     mycol3.delete_one(del_data)
     return jsonify({"message":"deleted shop"}),200
@@ -339,6 +417,9 @@ def del_shop(shop_id):
 #creating menu
 @app.route('/menu', methods=['POST'])
 def add_menu():
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     data = request.json
     shop_id = data["shop_id"]
     item_ids = data["item_ids"]
@@ -348,6 +429,9 @@ def add_menu():
 #editing menu
 @app.route('/menu/<shop_id>', methods=['PATCH'])
 def edit_menu(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     new_data = request.json
     item_ids = new_data["item_ids"]
     mycol4.update_one({"shop_id": shop_id}, { "$set" : new_data})
@@ -356,6 +440,9 @@ def edit_menu(shop_id):
 #empty menu
 @app.route('/menu/<shop_id>/delete', methods=['DELETE'])
 def empty_menu(shop_id):
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     del_data = { "shop_id" : shop_id }
     mycol4.delete_one(del_data)
     return jsonify({"message":"menu emptied"}),200
@@ -363,6 +450,9 @@ def empty_menu(shop_id):
 #generate order from cart
 @app.route('/cart/<customer_id>/order', methods=['GET'])
 def generate_order_by_id(customer_id):   
+    if 'token' not in session:
+            # If user have not login
+            return jsonify({"error":"you need to login first"}),401
     return jsonify({"message":"unfinished"}),200
